# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for the Kemit Trip Planner application.",
        "title": "Kemit Trip Planner API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/trips": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of trips for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "List user trips",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Trip"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates an AI-powered trip itinerary based on user input",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Create a new trip",
                "parameters": [
                    {
                        "description": "Trip creation parameters",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateTripInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Trip"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/trips/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get full details of a specific trip, including days and activities",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Get trip details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Trip ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Trip"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update details of an existing trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Update trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Trip update info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateTripInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Trip"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a trip and its associated data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Delete trip",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/trips/{id}/activities": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new activity to a trip",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Create activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.CreateActivityInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.TripActivity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
        "/trips/{id}/activities/{activityId}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Update activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID (UUID)",
                        "name": "activityId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity update info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateActivityInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.TripActivity"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an activity from a trip",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Delete activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID (UUID)",
                        "name": "activityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/trips/{id}/regenerate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Use AI to regenerate the trip itinerary",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trips"
                ],
                "summary": "Regenerate itinerary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trip ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Trip"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Preferences": {
            "type": "object",
            "properties": {
                "age_group": {
                    "description": "teen / adult / senior",
                    "type": "string"
                },
                "budget": {
                    "description": "low / medium / high",
                    "type": "string"
                },
                "crowd_preference": {
                    "description": "crowded / quiet / no_preference",
                    "type": "string"
                },
                "interests": {
                    "description": "[\"history\",\"adventure\",…]",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "season": {
                    "description": "winter / spring / summer / autumn",
                    "type": "string"
                },
                "travel_companion": {
                    "description": "solo / couple / family / friends",
                    "type": "string"
                }
            }
        },
...
        "models.Trip": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripDestination"
                    }
                },
                "end_date": {
                    "type": "string"
                },
                "expenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripExpense"
                    }
                },
                "id": {
                    "type": "string"
                },
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripMember"
                    }
                },
                "preferences": {
                    "description": "same JSON shape as user preferences",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Preferences"
                        }
                    ]
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "travelers_count": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.TripActivity": {
            "type": "object",
            "properties": {
                "activity_name": {
                    "type": "string"
                },
                "activity_type": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_hours": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "estimated_cost": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "order_in_day": {
                    "type": "integer"
                },
                "place": {
                    "$ref": "#/definitions/models.Place"
                },
                "place_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                },
                "trip_day": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TripDay"
                        }
                    ]
                },
                "trip_day_id": {
                    "type": "string"
                }
            }
        },
...
        "models.TripDestination": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "dominant category for this city",
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "days_allocated": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "order_in_trip": {
                    "type": "integer"
                },
                "trip": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Trip"
                        }
                    ]
                },
                "trip_days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TripDay"
                    }
                },
                "trip_id": {
                    "type": "string"
                }
            }
        },
        "models.TripExpense": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "trip_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.TripMember": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "role": {
                    "description": "owner, editor, viewer",
                    "type": "string"
                },
                "trip": {
                    "$ref": "#/definitions/models.Trip"
                },
                "trip_id": {
                    "type": "string"
                },
                "user": {
                    "description": "Relations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "services.CreateActivityInput": {
            "type": "object",
            "required": [
                "activity_name",
                "trip_day_id"
            ],
            "properties": {
                "activity_name": {
                    "type": "string"
                },
                "activity_type": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_hours": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "estimated_cost": {
                    "type": "number"
                },
                "place_id": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "trip_day_id": {
                    "type": "string"
                }
            }
        },
...
        "services.CreateTripInput": {
            "type": "object",
            "required": [
                "end_date",
                "preferences",
                "start_date",
                "travelers_count"
            ],
            "properties": {
                "end_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "preferences": {
                    "$ref": "#/definitions/models.Preferences"
                },
                "start_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "travelers_count": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "services.UpdateActivityInput": {
            "type": "object",
            "properties": {
                "activity_name": {
                    "type": "string"
                },
                "activity_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "duration_hours": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "estimated_cost": {
                    "type": "number"
                },
                "start_time": {
                    "type": "string"
                }
            }
        },
...
        "services.UpdateTripInput": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "travelers_count": {
                    "type": "integer"
                }
            }
        },
...
        "utils.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/utils.ErrorBody"
                },
                "meta": {
                    "$ref": "#/definitions/utils.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}